Updating the iterator data type and the README.md file.

+++ entro-hash.c
@@ -1,7 +1,7 @@
 #include "entro-hash.h"
 
 uint32_t entro_hash(const char *input, uint32_t entropy) {
-  unsigned long long i = 0;
+  unsigned long i = 0;
 
   while (input[i] != 0) {
     entropy ^= (input[i] + 111111) ^ 111111;
